web-backend/pyproject.toml
[project]
name = "barbershop-web-backend"
version = "0.1.0"
description = "Contact form / appointment booking backend for the barbershop website"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    # Web framework — [standard] bundles uvicorn, httpx, python-multipart, etc.
    "fastapi[standard]>=0.115.6",
    # Database
    "sqlalchemy[asyncio]>=2.0.0",
    "asyncpg>=0.30.0",
    "alembic>=1.14.0",
    "psycopg2-binary>=2.9.10",
    # Validation & Settings
    "pydantic>=2.10.4",
    "pydantic-settings>=2.7.0",
    "email-validator>=2.2.0",
    # Email
    "aiosmtplib>=3.0.2",
    # Utilities
    "python-dotenv>=1.0.1",
]

[tool.uv]
# Application — skip building/installing the project itself
package = false
